+++ Network/mASyncSerialPort.cpp
@@ -199,7 +199,7 @@ static void AsyncEvent( mASyncSerialPort& pipe , const mASyncSerialPort::NotifyO
 	}
 	else
 	{
-		RaiseAssert( g_ErrorLogger , 0 , L"非同期操作の完了通知方法が不正です" , info.Mode );
+		RaiseAssert( g_ErrorLogger , 0 , L"非同期操作の完了通知方法が不正です" , (int)info.Mode );
 	}
 }
